+++ inherited-rules/implicitly-included/eslint_conf_eslint-recommended.js
@@ -3,7 +3,10 @@ module.exports = {
   rules: {
     "for-direction": "error",
     "getter-return": "error",
+    "no-dupe-else-if": "error",
     "no-extra-semi": "error",
+    "no-import-assign": "error",
+    "no-setter-return": "error",
     "no-unused-labels": "error",
     "require-yield": "error"
   }